[이성준] 마지막 student 수정

+++ on-off/src/main/java/com/greedy/onoff/member/service/StudentManagerService.java
@@ -105,9 +105,10 @@ public class StudentManagerService {
 		
 		Pageable pageable = PageRequest.of(page - 1, 10, Sort.by("memberCode").descending());
 		
-		Page<Member> memberList = studentManagerRepository.findByMemberNameContainsAndMemberRole(pageable, memberName, "ROLE_STUDENT");		Page<MemberDto> memberDtoList = memberList.map(member -> modelMapper.map(member, MemberDto.class));
+		Page<Member> memberList = studentManagerRepository.findByMemberNameContainsAndMemberRole(pageable, memberName, "ROLE_STUDENT");		
+		Page<MemberDto> memberDtoList = memberList.map(member -> modelMapper.map(member, MemberDto.class));
 		/* 클라이언트 측에서 서버에 저장 된 이미지 요청 시 필요한 주소로 가공 */
-		memberDtoList.forEach(product -> product.setMemberImageUrl(IMAGE_URL + product.getMemberImageUrl()));
+		memberDtoList.forEach(student -> student.setMemberImageUrl(IMAGE_URL + student.getMemberImageUrl()));
 		
 		log.info("[memberList] memberDtoList : {}", memberDtoList.getContent());
 		
@@ -137,8 +138,10 @@ public class StudentManagerService {
 				replaceFileName = FileUploadUtils.saveFile(IMAGE_DIR, imageName, memberDto.getMemberImage());
 				memberDto.setMemberImageUrl(replaceFileName);
 				
-				/* 기존에 저장 된 이미지 삭제*/
-				FileUploadUtils.deleteFile(IMAGE_DIR, oriImage);
+				if(oriImage != null) {
+					/* 기존에 저장 된 이미지 삭제*/
+					FileUploadUtils.deleteFile(IMAGE_DIR, oriImage);
+				}
 
 			} else { 
 				/* 이미지를 변경하지 않는 경우 */
@@ -148,11 +151,11 @@ public class StudentManagerService {
 			oriStudent.studentUpdate(
 					memberDto.getMemberId(), 
 					memberDto.getMemberName(), 
+					memberDto.getMemberBirthday(),
+					memberDto.getMemberGender(),
+					memberDto.getMemberEmail(), 
 					memberDto.getMemberPhone(),  
 					memberDto.getMemberAddress(), 
-					memberDto.getMemberStatus(),
-					memberDto.getMemberEmail(), 
-					memberDto.getMemberRole(),
 					memberDto.getMemberImageUrl());
 			
 			studentManagerRepository.save(oriStudent);
@@ -176,6 +179,7 @@ public class StudentManagerService {
 		
 		log.info("[StudentService] signupStudent Start ============================");
 		log.info("[StudentService] memberDto : {}" + memberDto);
+		String replaceFileName = null;
 		
 		if(studentManagerRepository.findByMemberId(memberDto.getMemberId()) != null) {
 			log.info("[StudentService] 아이디가 중복 됩니다.");
@@ -187,13 +191,29 @@ public class StudentManagerService {
 			throw new DuplicateMemberEmailException("이메일이 중복 됩니다.");
 		}
 		
-		
-		memberDto.setMemberPassword(passwordEncoder.encode(memberDto.getMemberPassword()));
-		memberDto.setMemberRole("ROLE_STUDENT");
-		studentManagerRepository.save(modelMapper.map(memberDto, Member.class));
-		
-		
-		
+		try {
+			
+			if(memberDto.getMemberImage() != null) {
+				String imageName = UUID.randomUUID().toString().replace("-", "");
+				replaceFileName = FileUploadUtils.saveFile(IMAGE_DIR, imageName, memberDto.getMemberImage());
+				memberDto.setMemberImageUrl(replaceFileName);
+			}
+			
+			log.info("[StudentService] replaceFileName : {}", replaceFileName);
+			
+			
+			memberDto.setMemberPassword(passwordEncoder.encode(memberDto.getMemberPassword()));
+			memberDto.setMemberRole("ROLE_STUDENT");
+			studentManagerRepository.save(modelMapper.map(memberDto, Member.class));
+	
+		} catch (IOException e) {
+			e.printStackTrace();
+			 try {
+				 FileUploadUtils.deleteFile(IMAGE_DIR, replaceFileName);
+			} catch (IOException e1) {
+				e1.printStackTrace();
+			}	 
+		}
 		
 		log.info("[StudentService] signupStudent End ============================");
 		return memberDto;
@@ -232,6 +252,17 @@ public class StudentManagerService {
 	}
 	
 
+	/* 아이디 중복 체크 */
+	public Boolean checkMemberIdDuplicate(String memberId) {
+		log.info("[StudentService] checkMemberIdDuplicate Start ============================");
+
+		return studentManagerRepository.existsByMemberId(memberId);
+	}
+
+	/* 이메일 중복 확인 */
+	public Boolean checkMemberEmailDuplicate(String memberEmail) {
+		return studentManagerRepository.existsByMemberEmail(memberEmail);
+	}
